pyproject.toml
[project]
name = "slither-dataflow"
version = "0.1.0"
description = "Data flow analysis tools for Solidity smart contracts using Slither"
authors = [
    {name = "desp4c1to",email = "yyuannan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.8"

[tool.poetry.dependencies]
python = ">=3.10,<3.12.0"
slither-analyzer = "^0.11.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
slither-dataflow = "slither_dataflow.cli:main"